Move anchorRef back to component and install informed peer dependency
